requirements_bak.txt
accelerate                0.27.2
aiofiles                  23.2.1
aiohappyeyeballs          2.6.1
aiohttp                   3.12.0
aiosignal                 1.3.2
albucore                  0.0.24
albumentations            2.0.7
altair                    5.5.0
annotated-types           0.7.0
anthropic                 0.52.0
anyio                     3.7.1
astor                     0.8.1
attrs                     25.3.0
azure-core                1.34.0
azure-identity            1.23.0
beautifulsoup4            4.13.4
bitsandbytes              0.46.0
blinker                   1.9.0
boto3                     1.38.23
botocore                  1.38.23
cachetools                5.5.2
certifi                   2025.4.26
cffi                      1.17.1
cfgv                      3.4.0
charset-normalizer        3.4.2
click                     8.1.8
comtypes                  1.4.11
contourpy                 1.3.2
cryptography              45.0.3
cycler                    0.12.1
Cython                    3.1.1
dashscope                 1.23.3
decorator                 5.2.1
defusedxml                0.7.1
dill                      0.4.0
distlib                   0.3.9
distro                    1.9.0
easyocr                   1.7.2
einops                    0.8.0
fastapi                   0.115.12
ffmpy                     0.5.0
filelock                  3.18.0
fire                      0.7.0
flash_attn                2.7.4.post1
fonttools                 4.58.0
frozenlist                1.6.0
fsspec                    2025.5.1
gitdb                     4.0.12
GitPython                 3.1.44
google-auth               2.40.2
gradio                    5.13.2
gradio_client             1.6.0
groq                      0.25.0
h11                       0.16.0
hf-xet                    1.1.2
httpcore                  1.0.9
httpx                     0.28.1
huggingface-hub           0.32.1
identify                  2.6.12
idna                      3.10
imageio                   2.37.0
iniconfig                 2.1.0
Jinja2                    3.1.6
jiter                     0.10.0
jmespath                  1.0.1
jsonschema                4.22.0
jsonschema-specifications 2025.4.1
kiwisolver                1.4.8
lazy_loader               0.4
lmdb                      1.6.2
lxml                      5.4.0
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.10.3
mdurl                     0.1.2
MouseInfo                 0.1.3
mpmath                    1.3.0
msal                      1.32.3
msal-extensions           1.3.1
multidict                 6.4.4
narwhals                  1.40.0
networkx                  3.4.2
ninja                     1.11.1.4
nodeenv                   1.9.1
numpy                     2.1.1
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         9.1.0.70
nvidia-cufft-cu12         11.0.2.54
nvidia-cufile-cu12        1.11.1.6
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-cusparselt-cu12    0.6.3
nvidia-nccl-cu12          2.20.5
nvidia-nvjitlink-cu12     12.6.85
nvidia-nvtx-cu12          12.1.105
openai                    1.3.5
opencv-contrib-python     4.11.0.86
opencv-python             4.11.0.86
opencv-python-headless    4.11.0.86
opt-einsum                3.3.0
orjson                    3.10.18
packaging                 24.0
paddleocr                 2.10.0
paddlepaddle              3.0.0
pandas                    2.2.3
peft                      0.10.0
pillow                    11.2.1
pip                       25.1.1
platformdirs              4.3.8
pluggy                    1.6.0
pre-commit                3.8.0
propcache                 0.3.1
protobuf                  6.31.0
psutil                    7.0.0
py-cpuinfo                9.0.0
pyarrow                   20.0.0
pyasn1                    0.6.1
pyasn1_modules            0.4.2
PyAutoGUI                 0.9.54
pyclipper                 1.3.0.post6
pycparser                 2.22
pydantic                  2.10.6
pydantic_core             2.27.2
pydeck                    0.9.1
pydub                     0.25.1
PyGetWindow               0.0.9
Pygments                  2.19.1
PyJWT                     2.10.1
PyMsgBox                  1.0.9
pyparsing                 3.2.3
pyperclip                 1.9.0
PyRect                    0.2.0
PyScreeze                 1.0.1
pytest                    8.3.3
pytest-asyncio            0.23.6
python-bidi               0.6.6
python-dateutil           2.9.0.post0
python-docx               1.1.2
python-multipart          0.0.20
python3-xlib              0.15
pytweening                1.2.0
pytz                      2025.2
PyYAML                    6.0.2
RapidFuzz                 3.13.0
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
rich                      14.0.0
rpds-py                   0.25.1
rsa                       4.9.1
ruff                      0.6.7
s3transfer                0.13.0
safehttpx                 0.1.6
safetensors               0.5.3
scikit-image              0.25.2
scipy                     1.15.3
screeninfo                0.8.1
seaborn                   0.13.2
semantic-version          2.10.0
setuptools                80.8.0
shapely                   2.1.1
shellingham               1.5.4
simsimd                   6.2.1
six                       1.17.0
smmap                     5.0.2
sniffio                   1.3.1
soupsieve                 2.7
starlette                 0.46.2
streamlit                 1.45.1
stringzilla               3.12.5
supervision               0.18.0
sympy                     1.14.0
tenacity                  9.1.2
termcolor                 3.1.0
tifffile                  2025.5.24
timm                      1.0.15
tokenizers                0.15.2
toml                      0.10.2
tomlkit                   0.13.2
torch                     2.4.1+cu121
torchaudio                2.4.1+cu121
torchvision               0.19.1+cu121
tornado                   6.5.1
tqdm                      4.67.1
transformers              4.39.0
triton                    3.0.0        1
typer                     0.15.4
typing_extensions         4.13.2
typing-inspection         0.4.1
tzdata                    2025.2
uiautomation              2.0.28
ultralytics               8.3.70
ultralytics-thop          2.0.14
urllib3                   2.4.0
uvicorn                   0.34.2
virtualenv                20.31.2
watchdog                  6.0.0
websocket-client          1.8.0
websockets                14.2
wheel                     0.45.1
yarl                      1.20.0